refactor(timelines): add folder for timeline items
