+++ react-app/src/components/timelines/index.js
@@ -1,7 +1,7 @@
 
 export {default as Timelines} from './timelines.jsx';
-export {default as LikeItem} from './like-item.jsx';
-export {default as ThoughtItem} from './thought-item.jsx';
-export {default as ArticleItem} from './article-item.jsx';
-export {default as ProjectItem} from './project-item.jsx';
-export {default as CommentItem} from './comment-item.jsx';
+export {default as LikeItem} from './items/like-item.jsx';
+export {default as ThoughtItem} from './items/thought-item.jsx';
+export {default as ArticleItem} from './items/article-item.jsx';
+export {default as ProjectItem} from './items/project-item.jsx';
+export {default as CommentItem} from './items/comment-item.jsx';
